consistent example names
