+++ examples/run_Cp2kMultistageWorkChain_H2O.py
@@ -1,41 +1,32 @@
 # -*- coding: utf-8 -*-
-""" Test/example for the Cp2kMultistageWorkChain"""
+""" Test/example for the Cp2kMultistageWorkChain"""  # pylint: disable=invalid-name
 
 import sys
 import click
 import ase.build
 
 from aiida.engine import run
-from aiida.orm import Code, Dict, StructureData, Str, Int
+from aiida.orm import Code, StructureData, Str
 from aiida.common import NotExistent
 from aiida.plugins import WorkflowFactory
 
 Cp2kMultistageWorkChain = WorkflowFactory('lsmo.cp2k_multistage')  # pylint: disable=invalid-name
 
 
-def run_multistage_h2o_fail(cp2k_code):
+def run_multistage_h2o(cp2k_code):
     """Example usage: verdi run thistest.py cp2k@localhost"""
 
-    print('Testing CP2K multistage workchain on H2O')
-    print('>>> Making it fail because of an unphysical Multiplicity')
+    print('Testing CP2K multistage workchain on H2O (RKS, no need for smearing)...')
+    print(">>> Using 'test' tag")
 
     atoms = ase.build.molecule('H2O')
     atoms.center(vacuum=2.0)
     structure = StructureData(ase=atoms)
 
-    parameters = Dict(dict={'FORCE_EVAL': {
-        'DFT': {
-            'UKS': True,
-            'MULTIPLICITY': 666,
-        }
-    }})
-
     # Construct process builder
     builder = Cp2kMultistageWorkChain.get_builder()
     builder.structure = structure
     builder.protocol_tag = Str('test')
-    builder.cp2k_base.max_iterations = Int(1)
-    builder.cp2k_base.cp2k.parameters = parameters
     builder.cp2k_base.cp2k.code = cp2k_code
     builder.cp2k_base.cp2k.metadata.options.resources = {
         'num_machines': 1,
@@ -55,7 +46,7 @@ def cli(codelabel):
     except NotExistent:
         print("The code '{}' does not exist".format(codelabel))
         sys.exit(1)
-    run_multistage_h2o_fail(code)
+    run_multistage_h2o(code)
 
 
 if __name__ == '__main__':
