+++ examples/run_Cp2kMultistageWorkChain_H2O_test_file.py
@@ -1,32 +1,36 @@
 # -*- coding: utf-8 -*-
-""" Test/example for the Cp2kMultistageWorkChain"""
+""" Test/example for the Cp2kMultistageWorkChain"""  # pylint: disable=invalid-name
 
+import os
 import sys
 import click
 import ase.build
 
 from aiida.engine import run
-from aiida.orm import Code, StructureData, Str
+from aiida.orm import Code, StructureData, SinglefileData
 from aiida.common import NotExistent
 from aiida.plugins import WorkflowFactory
 
 Cp2kMultistageWorkChain = WorkflowFactory('lsmo.cp2k_multistage')  # pylint: disable=invalid-name
 
 
-def run_multistage_h2o_singlepoint(cp2k_code):
+def run_multistage_h2o_testfile(cp2k_code):
     """Example usage: verdi run thistest.py cp2k@localhost"""
 
     print('Testing CP2K multistage workchain on H2O')
-    print(">>> Using 'singlepoint' tag, with no output structure")
+    print('>>> Loading a custom protocol from file testfile.yaml')
 
     atoms = ase.build.molecule('H2O')
     atoms.center(vacuum=2.0)
     structure = StructureData(ase=atoms)
 
+    thisdir = os.path.dirname(os.path.abspath(__file__))
+
     # Construct process builder
     builder = Cp2kMultistageWorkChain.get_builder()
     builder.structure = structure
-    builder.protocol_tag = Str('singlepoint')
+    builder.protocol_yaml = SinglefileData(
+        file=os.path.abspath(os.path.join(thisdir, '..', 'data', 'test_multistage_protocol.yaml')))
     builder.cp2k_base.cp2k.code = cp2k_code
     builder.cp2k_base.cp2k.metadata.options.resources = {
         'num_machines': 1,
@@ -46,7 +50,7 @@ def cli(codelabel):
     except NotExistent:
         print("The code '{}' does not exist".format(codelabel))
         sys.exit(1)
-    run_multistage_h2o_singlepoint(code)
+    run_multistage_h2o_testfile(code)
 
 
 if __name__ == '__main__':
