+++ examples/test_IsothermInflectionWorkChain_graphite_Ar.py
